Still not working, much better though
The system is still not working, but it is doing something, there is now
just a lot of debugging to do in order to get it working. The LFSR has
been implemented as a program counter and it appears to be doing the
right thing, all the pieces are here.